refactor(tau-session): extract session store runtime module
